+++ src/test/java/org/corpus_tools/annis/gui/HelperTest.java
@@ -5,8 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Arrays;
-import java.util.List;
-import org.corpus_tools.annis.gui.Helper;
 import org.junit.jupiter.api.Test;
 
 class HelperTest {
@@ -30,24 +28,23 @@ class HelperTest {
 
   @Test
   void testBuildAQLDocumentQuery() {
-    List<String> docPath = Arrays.asList("corpus", "doc");
     // Test the full document fallback query
     assertEquals(WHOLEDOCUMENT_QUERY,
-        Helper.buildDocumentQuery(docPath, null, false));
+        Helper.buildDocumentQuery("corpus/doc", null, false));
     // Test the raw text query
     assertEquals(RAWTEXT_QUERY,
-        Helper.buildDocumentQuery(docPath, null, true));
+        Helper.buildDocumentQuery("corpus/doc", null, true));
     // Test the node annotation filter
     assertEquals(
         "(a#tok | a#pos | a#something) & doc#annis:node_name=/corpus\\x2Fdoc/ & #a @* #doc",
-        Helper.buildDocumentQuery(docPath, Arrays.asList("pos", "something"), true));
+        Helper.buildDocumentQuery("corpus/doc", Arrays.asList("pos", "something"), true));
    // Check invalid node annotation names lead to falling back to the whole document query
    assertEquals(WHOLEDOCUMENT_QUERY,
-       Helper.buildDocumentQuery(docPath, Arrays.asList("pos", "myanno=something"), false));
+       Helper.buildDocumentQuery("corpus/doc", Arrays.asList("pos", "myanno=something"), false));
    // When giving giving both a node annotation filter argument and using raw text, the node filter
    // query should be returned
    assertEquals("(a#tok | a#ns:pos) & doc#annis:node_name=/corpus\\x2Fdoc/ & #a @* #doc",
-       Helper.buildDocumentQuery(docPath, Arrays.asList("ns:pos"), true));
+       Helper.buildDocumentQuery("corpus/doc", Arrays.asList("ns:pos"), true));
 
     
   }