+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Documents which special characters (like %) in their name could not be opened
+  in full text visualizes (like the document browser).
+- The `raw_text` visualizer did not work properly and did not show the whole
+  text.
 - Avoid issues with possible problematic random number generation in visualizer
   result ID. Since the ID is only locally used for one match view, the impact of
   a flawed random generator are limited, but it is still bad practice to create